+++ export/config/current/sheet_configs/config_collett_tronchi.json
@@ -151,23 +151,6 @@
                         "else": 0
                     }
                 }
-            },"validations": [{
-                    "func": "IF",
-                    "params": {
-                        "field": "102000",
-                        "cond": [{
-                            "and": [
-                              {"lookup": "{68600}", "operator": "=", "value": 1},
-                              {"operator":  "=", "value": 0}
-                            ]
-                        },{
-                          "and": [
-                              {"lookup": "{68600}", "operator": "!=", "value": 1}
-                            ]
-                        }]
-                    },
-                    "warning": "Foglio: {SHEET}|Riga: {ROW}|Codice_ato: {CODICE_ATO}|Campo: {FIELD}|se a parità di codice ATO il campo ID 68600 = 1, allora il campo assume valore 0"
-                }
-            ]}
+            }}
      ]
 }
